+++ assets/js/collaborative-editor/components/CollaborativeJobEditor.tsx
@@ -15,7 +15,7 @@ import Docs from "../../adaptor-docs/Docs";
 import { Tabs } from "../../components/Tabs";
 import Metadata from "../../metadata-explorer/Explorer";
 import { useAwarenessReady, useRawAwareness } from "../hooks/useAwareness";
-import { useWorkflowSelector } from "../hooks/useWorkflow";
+import { useWorkflowSelector, useWorkflowReadOnly } from "../hooks/useWorkflow";
 
 import { CollaborativeMonaco } from "./CollaborativeMonaco";
 import { CollaborativeWorkflowDiagram } from "./diagram/CollaborativeWorkflowDiagram";
@@ -64,6 +64,7 @@ export function CollaborativeJobEditor({
   const awareness = useRawAwareness();
   const awarenessReady = useAwarenessReady();
   logger.log("awarenessReady", awarenessReady);
+  const { isReadOnly } = useWorkflowReadOnly();
 
   const [vertical, setVertical] = useState(
     () => settings[SettingsKeys.ORIENTATION] === "v"
@@ -125,7 +126,7 @@ export function CollaborativeJobEditor({
               ytext={jobBodyYText!}
               awareness={awareness!}
               adaptor={adaptor}
-              disabled={disabled}
+              disabled={isReadOnly}
               className="h-full w-full"
             />
             {disabled && disabledMessage && (